tests/wpt/web-platform-tests/keyboard-lock/navigator-keyboard-unlock.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

test(() => {
  assert_equals(navigator.keyboard.unlock(),
                undefined);
}, '[Keyboard Lock] keyboard.unlock');

</script>
